refactor: extract shared CodeBlock component from duplicated highlighters
JsonPreview, PythonPreview, and TextPreview in AppCodePreview.tsx each
had a near-identical SyntaxHighlighter block (same style selection,
customStyle, wrap options), and MarkdownText's pre renderer had just
added a fourth copy of the same pattern. Extracted a shared CodeBlock
component that owns the isDark lookup and style config, removing the
now-redundant isDark prop threading through AppCodePreview's preview
components and helper function.
